system.py: Added system class
The instances of this class are the main feature of this program
as it contains the full description of the system. The rest of the
elements of this package are subroutines to populate and check this
class, which when completed should provide a system ready for
simulation.

This class is loaded by default when the package is imported.